Extract app/ layer; thin main.dart to its single job
The architecture doc (lib/context/architecture.context.md section 3)
prescribes a three-layer lib/ structure (app/, core/, features/) where
the app shell lives under app/ and main.dart's "only job is to
initialize bindings and run the app". The codebase was missing the
app/ layer entirely - main.dart was doing three jobs: initialization,
hosting the MaterialApp, and routing.

- New lib/app/app.dart holds TurboApp, the MaterialApp widget
  responsible for theme selection, localization, and routing.
  Renamed from MyApp.
- New lib/app/app_theme.dart consolidates the MaterialTheme class
  (color schemes) and the createTextTheme helper. Moved from
  lib/core/theme/ - they were only used in main.dart and are
  app-shell concerns per the doc.
- lib/l10n/app_localizations.dart moves to lib/app/l10n/ per the doc.
  All 40+ imports across lib/ and test/ updated to the new path.
- lib/main.dart is now 43 lines (was 84): just bindings, container
  creation, background-init kickoff, and runApp.
- lib/core/theme/ removed.

The _kickOffBackgroundInit function stays in main.dart since it's
literally part of initialization - it just triggers async feature
providers without blocking the first frame.

Verified: flutter analyze clean, all 129 tests pass (including the
two architecture boundary tests).
