+++ lib/features/auth/widgets/dev_banner.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 
 /// A banner to indicate when the app is in development mode
 class DevModeBanner extends StatelessWidget {
