+++ lib/features/auth/widgets/drawer_widget.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:turbo/features/settings/api.dart';
 import 'package:turbo/features/tile_storage/offline_regions/api.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import './user_profile_screen.dart';
 import './login_screen.dart';
 import '../data/auth_providers.dart';
