+++ lib/features/auth/widgets/google_oauth_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'dart:async';
 
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import '../data/auth_providers.dart';
 
 class GoogleAuthCallbackPage extends ConsumerStatefulWidget {
