+++ lib/features/auth/widgets/google_sign_in_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:turbo/core/config/env_config.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import 'package:url_launcher/url_launcher.dart';
 import '../data/auth_providers.dart';
