+++ lib/features/auth/widgets/login_success.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import '../data/auth_providers.dart';
 
 class LoginSuccessPage extends ConsumerStatefulWidget {
