+++ lib/features/auth/widgets/register_view_mobile.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../data/auth_providers.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import './auth_divider.dart';
 import './auth_error_message.dart';
 import './auth_footer_link.dart';
