+++ lib/features/markers/widgets/edit_location_sheet.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import '../data/icon_service.dart';
 import '../models/marker.dart';
 import '../models/named_icon.dart';
