+++ lib/features/measuring/widgets/measuring_controls.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 
 class MeasuringControls extends StatelessWidget {
   final double distance;
