+++ lib/features/measuring/widgets/measuring_map_page.dart
@@ -17,7 +17,7 @@ import 'package:turbo/core/widgets/map/controls/bottom_controls.dart';
 import 'package:turbo/core/widgets/map/controls/default_map_controls.dart';
 import 'package:turbo/core/widgets/map/controls/go_back_button.dart';
 import 'package:turbo/core/widgets/map/controls/map_controls.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 
 import '../../map_view/api.dart';
 import '../../tile_providers/api.dart';
