+++ lib/features/saved_paths/widgets/export_options_sheet.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 
 import '../data/path_export_service.dart';
 import '../models/saved_path.dart';
