+++ lib/features/saved_paths/widgets/path_detail_sheet.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import 'package:turbo/core/widgets/buttons/primary_button.dart';
 import '../models/saved_path.dart';
 import '../models/path_style.dart';
