+++ lib/features/saved_paths/widgets/path_info_sheet.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:turbo/core/widgets/action_button.dart';
 import 'package:turbo/features/markers/api.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 import '../data/saved_path_repository.dart';
 import '../models/path_style.dart';
 import '../models/saved_path.dart';
