+++ lib/features/tile_providers/data/providers/norges_kart_topo.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:turbo/features/tile_providers/models/tile_provider_config.dart';
-import 'package:turbo/l10n/app_localizations.dart';
+import 'package:turbo/app/l10n/app_localizations.dart';
 
 class NorgeskartTopoConfig extends TileProviderConfig {
   @override
